Various changes
Made changes both to front-end and backend so that a user can log-in and log-out. At the same time the home page for the user is set to the dashboard, as long as he is logged in, for that purpose created a home.jsx page that will handle the toggling between pages.