Finalized the app
A lot of changes both to back and front end of the app to create the current conversation window. made changes to the back end to my serializers- views-urls to create new end point to provide the necessary results from the front end queries. made changes to the front end of the nav bar , created 2 files for the nav bars . The first is for unauthorized users and the other for authorized ones . created a 404 page and created some new query hooks to help with fast response and queries needed from back-end.